Mutate new columns

Applies the ballet() and makes sure each result have the same vctrs::vec_size() as the number of elements in each group.

salsa(.tbl, ..., .env = caller_env())

chacha(.tbl)

samba(.tbl, ..., .env = caller_env())

madison(.tbl, ..., .name = "data", .env = caller_env())

Arguments

.tbl
A data frame, most likely a grouped data frame

..., .env
formulas for each column to create, and parent environment, see ballet()

.name
Name of the packed column made by charleston()

The four functions play a separate role around the idea of dplyr::mutate():

chacha() reorganizes the rows of a grouped data frame so that data for each group is contiguous in each column.

salsa() runs the ballet() defined by ... and makes sure the vctrs::vec_size() of each result is equal to the number of elements in that group. The result tibble of salsa() does not contain the grouping variables, just those columns specified by the ....

samba() is the closest to dplyr::mutate(), it column binds the result of chacha() and salsa() with vctrs::vec_cbind().

madison() is similar to sambda() but the results are packed instead of being [vctrs::vec_cbind()]. The name of the created packed column is controled by the .name argument.

Examples

g <- group_by(iris, Species)

# Creates a `dance_grouped_df` tibble,
# which is the same as `g` but guarantees that the data for each column
# is contiguous within groups
chacha(g)

#> # A tibble: 150 x 5
#> # Groups:   Species [3]
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#>  1          5.1         3.5          1.4         0.2 setosa 
#>  2          4.9         3            1.4         0.2 setosa 
#>  3          4.7         3.2          1.3         0.2 setosa 
#>  4          4.6         3.1          1.5         0.2 setosa 
#>  5          5           3.6          1.4         0.2 setosa 
#>  6          5.4         3.9          1.7         0.4 setosa 
#>  7          4.6         3.4          1.4         0.3 setosa 
#>  8          5           3.4          1.5         0.2 setosa 
#>  9          4.4         2.9          1.4         0.2 setosa 
#> 10          4.9         3.1          1.5         0.1 setosa 
#> # … with 140 more rows

# returns a tibble of two columns
g %>%
  salsa(
    Sepal = ~Sepal.Length * Sepal.Width,
    Petal = ~Petal.Length * Petal.Width
  )

#> # A tibble: 150 x 2
#>    Sepal Petal
#>    <dbl> <dbl>
#>  1  17.8 0.280
#>  2  14.7 0.280
#>  3  15.0 0.26 
#>  4  14.3 0.3  
#>  5  18   0.280
#>  6  21.1 0.68 
#>  7  15.6 0.42 
#>  8  17   0.3  
#>  9  12.8 0.280
#> 10  15.2 0.15 
#> # … with 140 more rows

# returns a dance_grouped_df with the two
# additional columns `Sepal` and `Petal`
g %>%
  samba(
    Sepal = ~Sepal.Length * Sepal.Width,
    Petal = ~Petal.Length * Petal.Width
  )

#> # A tibble: 150 x 7
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species Sepal Petal
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>   <dbl> <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa   17.8 0.280
#>  2          4.9         3            1.4         0.2 setosa   14.7 0.280
#>  3          4.7         3.2          1.3         0.2 setosa   15.0 0.26 
#>  4          4.6         3.1          1.5         0.2 setosa   14.3 0.3  
#>  5          5           3.6          1.4         0.2 setosa   18   0.280
#>  6          5.4         3.9          1.7         0.4 setosa   21.1 0.68 
#>  7          4.6         3.4          1.4         0.3 setosa   15.6 0.42 
#>  8          5           3.4          1.5         0.2 setosa   17   0.3  
#>  9          4.4         2.9          1.4         0.2 setosa   12.8 0.280
#> 10          4.9         3.1          1.5         0.1 setosa   15.2 0.15 
#> # … with 140 more rows

# returns a dance_grouped_df with the one
# additional data frame column
g %>%
  madison(
    Sepal = ~Sepal.Length * Sepal.Width,
    Petal = ~Petal.Length * Petal.Width
  )

#> # A tibble: 150 x 6
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species data$Sepal $Petal
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>        <dbl>  <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa        17.8  0.280
#>  2          4.9         3            1.4         0.2 setosa        14.7  0.280
#>  3          4.7         3.2          1.3         0.2 setosa        15.0  0.26 
#>  4          4.6         3.1          1.5         0.2 setosa        14.3  0.3  
#>  5          5           3.6          1.4         0.2 setosa        18    0.280
#>  6          5.4         3.9          1.7         0.4 setosa        21.1  0.68 
#>  7          4.6         3.4          1.4         0.3 setosa        15.6  0.42 
#>  8          5           3.4          1.5         0.2 setosa        17    0.3  
#>  9          4.4         2.9          1.4         0.2 setosa        12.8  0.280
#> 10          4.9         3.1          1.5         0.1 setosa        15.2  0.15 
#> # … with 140 more rows