Apply a single function to multiple columns

swing() : returns a spliced list of formulas, suitable for the ... argument of choreography(), as well as all the dances that use a choreography()

swing(.fun, ..., .tbl = get_tbl(), .name = "{var}",
  .env = caller_env())

twist(.fun, ..., .tbl = get_tbl(), .name = "{var}",
  .env = caller_env())

Arguments

.fun
A function or a formula that uses . or .x to refer to each of the selected column

...
tidy selection of columns, see tidyselect::vars_select() for details

.tbl, .env
data frame ... selects columns from, this is automatically set by the choreography(), you should rarely need to use these arguments

.name
glue::glue() model to name the outputs. The model may use :

{var} to refer to the name of the current selected variable

{idx} to refer to the index of the current variable The default value "var" for .name simply uses the name of the selected variable

Details

twist() : returns a single formula that makes a tibble column

These functions are generally used within other dances, such as tango(), samba() or jive()

See also

Examples

g <- iris %>% group_by(Species)

##------- tango()

# Apply mean to all columns that start with Sepal
# and choose how the result columns are called
g %>%
  tango(
    swing(mean, starts_with("Sepal"), .name = "mean_{var}")
  )

#> # A tibble: 3 x 3
#>   Species    mean_Sepal.Length mean_Sepal.Width
#>   <fct>                  <dbl>            <dbl>
#> 1 setosa                  5.01             3.43
#> 2 versicolor              5.94             2.77
#> 3 virginica               6.59             2.97

# if you want to use extra arguments of `.fun` you can embed
# them with the lambda syntax
g %>%
  tango(
    swing(~mean(., trim = .2), starts_with("Sepal"), .name = "mean_{var}")
  )

#> # A tibble: 3 x 3
#>   Species    mean_Sepal.Length mean_Sepal.Width
#>   <fct>                  <dbl>            <dbl>
#> 1 setosa                  5                3.41
#> 2 versicolor              5.91             2.80
#> 3 virginica               6.55             2.96

# use twist() to instead create a single packed column
g %>%
  tango(
    mean = twist(mean, starts_with("Sepal"))
  )

#> # A tibble: 3 x 2
#>   Species    mean$Sepal.Length $Sepal.Width
#>   <fct>                  <dbl>        <dbl>
#> 1 setosa                  5.01         3.43
#> 2 versicolor              5.94         2.77
#> 3 virginica               6.59         2.97

# but in fact, if you don't name the formula made by twist()
# the columns are auto spliced
g %>%
  tango(
    twist(mean, starts_with("Sepal"))
  )

#> # A tibble: 3 x 3
#>   Species    Sepal.Length Sepal.Width
#>   <fct>             <dbl>       <dbl>
#> 1 setosa             5.01        3.43
#> 2 versicolor         5.94        2.77
#> 3 virginica          6.59        2.97

##------- samba()

g %>%
  samba(
    swing(~. - mean(.), starts_with("Sepal"), .name = "centered_{var}")
  )

#> # A tibble: 150 x 7
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species centered_Sepal.…
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>              <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa            0.0940
#>  2          4.9         3            1.4         0.2 setosa           -0.106 
#>  3          4.7         3.2          1.3         0.2 setosa           -0.306 
#>  4          4.6         3.1          1.5         0.2 setosa           -0.406 
#>  5          5           3.6          1.4         0.2 setosa           -0.006 
#>  6          5.4         3.9          1.7         0.4 setosa            0.394 
#>  7          4.6         3.4          1.4         0.3 setosa           -0.406 
#>  8          5           3.4          1.5         0.2 setosa           -0.006 
#>  9          4.4         2.9          1.4         0.2 setosa           -0.606 
#> 10          4.9         3.1          1.5         0.1 setosa           -0.106 
#> # … with 140 more rows, and 1 more variable: centered_Sepal.Width <dbl>

g %>%
  samba(
    centered = twist(~. - mean(.), starts_with("Sepal"), .name = "centered_{var}")
  )

#> # A tibble: 150 x 6
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species centered$center…
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>              <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa            0.0940
#>  2          4.9         3            1.4         0.2 setosa           -0.106 
#>  3          4.7         3.2          1.3         0.2 setosa           -0.306 
#>  4          4.6         3.1          1.5         0.2 setosa           -0.406 
#>  5          5           3.6          1.4         0.2 setosa           -0.006 
#>  6          5.4         3.9          1.7         0.4 setosa            0.394 
#>  7          4.6         3.4          1.4         0.3 setosa           -0.406 
#>  8          5           3.4          1.5         0.2 setosa           -0.006 
#>  9          4.4         2.9          1.4         0.2 setosa           -0.606 
#> 10          4.9         3.1          1.5         0.1 setosa           -0.106 
#> # … with 140 more rows, and 1 more variable: $centered_Sepal.Width <dbl>

##------- jive()

g %>%
  jive(
    q = ~ c("25%", "50%", "75%"),
    swing(~quantile(., c(0.25, 0.5, 0.75)), contains("."))
  )

#> # A tibble: 9 x 6
#> # Groups:   Species [3]
#>   Species    q     Sepal.Length Sepal.Width Petal.Length Petal.Width
#>   <fct>      <chr>        <dbl>       <dbl>        <dbl>       <dbl>
#> 1 setosa     25%           4.8         3.2          1.4          0.2
#> 2 setosa     50%           5           3.4          1.5          0.2
#> 3 setosa     75%           5.2         3.68         1.58         0.3
#> 4 versicolor 25%           5.6         2.52         4            1.2
#> 5 versicolor 50%           5.9         2.8          4.35         1.3
#> 6 versicolor 75%           6.3         3            4.6          1.5
#> 7 virginica  25%           6.22        2.8          5.1          1.8
#> 8 virginica  50%           6.5         3            5.55         2  
#> 9 virginica  75%           6.9         3.18         5.88         2.3